// projects/ui/src/lib/components/po-chart/interfaces/po-chart-axes.interface.ts
/**
 * @description
 *
 * *Interface* que define os eixos do grid.
 */
export interface PoChartAxes {
  /**
   * Número de linhas exibidas no eixo X.
   * - Valor padrão: '5';
   * - Valor mínimo permitido: '2';
   * - Máximo Máximo permitido: '10';
   */
  axisXGridLines?: number;

  /**
   * Define o alcance de valor máximo exibido no eixo Y.
   * Caso não seja definido valor, o valor de alcance máximo exibido será o maior existente entre as séries.
   * Se por acaso o valor máximo das séries for superior ao definido aqui, esta propriedade será ignorada.
   *
   * > Esta definição não deve refletir na plotagem das séries. Os valores máximos e mínimos encontrados nas séries serão as bases para seus alcance.
   */
  maxRange?: number;

  /**
   * Define o alcance mínimo exibido no eixo Y.
   * Caso não seja definido valor, o valor de alcance mínimo exibido será zero.
   * Se houver valores negativos nas séries, o menor deles será a base mínima.
   * Se por acaso o valor mínimo das séries for inferior ao definido aqui, esta propriedade será ignorada.
   *
   * > Esta definição não deve refletir na plotagem das séries. Os valores máximos e mínimos encontrados nas séries serão as bases para seus alcance.
   */
  minRange?: number;
}